move v3 tests that just GET data to test module that uses db fixture rather than fresh_db fixture
