+++ flexmeasures/api/v3_0/tests/test_sensor_data_fresh_db.py
@@ -1,12 +1,10 @@
 from __future__ import annotations
 
-from datetime import timedelta
-
 import pytest
 from flask import url_for
 from timely_beliefs.tests.utils import equal_lists
 
-from flexmeasures import Sensor, Source, User
+from flexmeasures import Sensor, Source
 from flexmeasures.api.tests.utils import get_auth_token
 from flexmeasures.api.v3_0.tests.utils import make_sensor_data_request_for_gas_sensor
 from flexmeasures.data.models.time_series import TimedBelief
@@ -116,67 +114,3 @@ def test_auto_fix_missing_registration_of_user_as_data_source(
         user=setup_user_without_data_source
     ).one_or_none()
     assert data_source is not None
-
-
-def test_get_sensor_data(
-    client,
-    setup_api_fresh_test_data: dict[str, Sensor],
-    setup_roles_users_fresh_db: dict[str, User],
-):
-    """Check the /sensors/data endpoint for fetching 1 hour of data of a 10-minute resolution sensor."""
-    sensor = setup_api_fresh_test_data["some gas sensor"]
-    source: Source = setup_roles_users_fresh_db["Test Supplier User"].data_source[0]
-    assert sensor.event_resolution == timedelta(minutes=10)
-    message = {
-        "sensor": f"ea1.2021-01.io.flexmeasures:fm1.{sensor.id}",
-        "start": "2021-05-02T00:00:00+02:00",
-        "duration": "PT1H20M",
-        "horizon": "PT0H",
-        "unit": "m³/h",
-        "source": source.id,
-        "resolution": "PT20M",
-    }
-    auth_token = get_auth_token(client, "test_supplier_user_4@seita.nl", "testtest")
-    response = client.get(
-        url_for("SensorAPI:get_data"),
-        query_string=message,
-        headers={"content-type": "application/json", "Authorization": auth_token},
-    )
-    print("Server responded with:\n%s" % response.json)
-    assert response.status_code == 200
-    values = response.json["values"]
-    # We expect two data points (from conftest) followed by 2 null values (which are converted to None by .json)
-    # The first data point averages [91.3, 91.7], and the second data point averages [92.1, None].
-    assert all(a == b for a, b in zip(values, [91.5, 92.1, None, None]))
-
-
-def test_get_instantaneous_sensor_data(
-    client,
-    setup_api_fresh_test_data: dict[str, Sensor],
-    setup_roles_users_fresh_db: dict[str, User],
-):
-    """Check the /sensors/data endpoint for fetching 1 hour of data of an instantaneous sensor."""
-    sensor = setup_api_fresh_test_data["some temperature sensor"]
-    source: Source = setup_roles_users_fresh_db["Test Supplier User"].data_source[0]
-    assert sensor.event_resolution == timedelta(minutes=0)
-    message = {
-        "sensor": f"ea1.2021-01.io.flexmeasures:fm1.{sensor.id}",
-        "start": "2021-05-02T00:00:00+02:00",
-        "duration": "PT1H20M",
-        "horizon": "PT0H",
-        "unit": "°C",
-        "source": source.id,
-        "resolution": "PT20M",
-    }
-    auth_token = get_auth_token(client, "test_supplier_user_4@seita.nl", "testtest")
-    response = client.get(
-        url_for("SensorAPI:get_data"),
-        query_string=message,
-        headers={"content-type": "application/json", "Authorization": auth_token},
-    )
-    print("Server responded with:\n%s" % response.json)
-    assert response.status_code == 200
-    values = response.json["values"]
-    # We expect two data point (from conftest) followed by 2 null values (which are converted to None by .json)
-    # The first data point is the first of [815, 817], and the second data point is the first of [818, None].
-    assert all(a == b for a, b in zip(values, [815, 818, None, None]))
